Add simple versioning mechanism to SLALite
The objective is to generate Docker images automatically tagged with the
current version (see below) and when the image is run (or SLALite
standalone), the logs show "Running SLALite vx.y.z"

resources/bin/release is a simple script that creates a git tag pointing
to the current branch and commit. It accepts a parameter (patch, minor
or major), automatically calculates the next version number from the
current version number (the current version is the last line on a `git
tag -l`), and tags the HEAD with the next version number. More complex
things can be done (like generating a CHANGELOG), but versionist already
exists for this.

Makefile contains targets:

* build: build SLALite trying to obtain a version identifier and
inserting into `main.version`. If the HEAD is tagged, it will use the
tag name. If not, a commit id.
* docker: create the Docker image, passing VERSION and DATE vars. The
Dockerfile builds the image using these vars, like build target does.
The image is tagged with the VERSION var.