Fix replace GetActiveAgreements with GetAgreementsByState
The new function is more generic and avoids creating several functions
for each possible state search.

IMPORTANT: Behavior has been modified. GetActiveAgreements internally
checked that the agreement was not expired (i.e. started agreements with
expired date were not returned), but this was not documented. The
consequence is that evaluation could never get agreements to set them as
terminated without adding extra logic. Evaluation is executed each
minute by default and getting expired but started agreements should not
be a problem as long as the agreement is not evaluated.