Make MonitoringAdapter API stateless
API was stateful. The idea of this was to perform an early retrieval
from monitoring (i.e., retrieve all metrics in one request).  This is
not a problem now, but it will be if agreements are evaluated in parallel
using goroutines. The simplest way to fix this is returning an initialized
copy of the MonitoringAdapter in Initialize().