Add generic adapter and most of the tests
Implements a generic adapter that acepts Retrieve and Process functions
to perform data retrieval from monitoring and additional processing
(like metric aggregation). The adapter mounts the data for the
assessment process into a GuaranteeData in a way that it is not needed
that values of the different series happen at the same times.

TODO: generic adapter tests